generatePlaylist endpoint completed
