+++ token.json
@@ -1,1 +1,1 @@
-{"access_token":"BQBlm2J04zPt37YibZUi4Gr_fmVx7jNehRDfGpLDpK5eWIEwKn6QTpDHmwlm8Dsn-UFUzA9JYDxNIUuj8-XZIzbd7GpJ9JMZZKD_BdGISRxC2xs_obZUCO2enWly3-MiHtrW2JadHdP7mhiVqnZomwHzfSzIjbtGGOAEsVLuvxDQuBGGUYXLvGtuUbruGkhzHMfSuG20LaNkeOcU_obd3e0k_TSsez_5KMQXaf-oJigexuRKLy8","token_type":"Bearer","expires_in":3600,"refresh_token":"AQA-uBqDh7A8oTVdUWJx3hdhHunytt4jt1HA3Z5bnTJyBbGBS7-5sSA8mxA0ckVX6ZQdcr2tDBZJAz1q79e0deUUEZcnHyVkOSJdJxk5lk5-MwMj2y42-GJ6KKk3sjyTeOA","scope":"playlist-modify-public user-read-email user-read-private"}
+{"access_token":"BQBXTdbtocM_2V675xKhpE1-2BNbluFx4MlV46-I_iJxmigSPXurSk6spOZo_5PXgEuTcgUO-eMzbE1MpBC968GyaiYzUxiQgeqHZWF-Ep-_KObAfDfOm5M3QQfUgaDUtQ0PMC5ysQh8nd9rGy6PORRP_Rh3k4LlsWMP4YRLU8o_9In3Cr3HhvzZ1vM0WjEILE98PoFORYV7bFhoNLrM8zN9eF7NfPpJ9T-0QvKRrwBDKOEh6FU","token_type":"Bearer","expires_in":3600,"refresh_token":"AQAyTYN1pK0apnxJRGPCt2-05RcbgPSwFAZRq1_ANN7Mj3RQd0sApj3oZxecCRG55axrr4-OMWh5pgLHNpkXwcZcJn3ueJCwi9vOZQan_SVcwGMm8sjl9Le2PpSl851_mjg","scope":"playlist-modify-public user-read-email user-read-private"}
